beginning manager part

+++ lib/features/customers/presentation/page/manager_part/widget/customer_list.dart
@@ -4,17 +4,19 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sharq_crm/features/customers/domain/entity/customer_entity.dart';
 
 import 'package:sharq_crm/features/customers/presentation/bloc/customer_cubit.dart';
+import 'package:sharq_crm/features/customers/presentation/page/manager_part/customers_page.dart';
 
 import '../customer_detail_page.dart';
+import '../customer_update_page.dart';
+
+enum ActionItems { delete, update }
 
 class CustomerList extends StatefulWidget {
   CustomerList({Key? key, required this.customersList})
       : super(
           key: key,
         );
-  TextEditingController nameController = TextEditingController();
 
-  TextEditingController phoneController = TextEditingController();
   List<CustomerEntity> customersList;
 
   @override
@@ -22,10 +24,12 @@ class CustomerList extends StatefulWidget {
 }
 
 class _CustomerListState extends State<CustomerList> {
+  ActionItems? selectedMenu;
+
   @override
   void setState(VoidCallback fn) {
-    context.read<CustomerCubit>().loadCustomer();
-    print('list ichida loadCustomer() chaqirildi.}');
+    // context.read<CustomerCubit>().loadCustomer();
+    // print('list ichida loadCustomer() chaqirildi.}');
     super.setState(fn);
   }
 
@@ -39,34 +43,52 @@ class _CustomerListState extends State<CustomerList> {
   Widget _listViewSeparated(
       List<CustomerEntity> customersList, BuildContext context) {
     return ListView.separated(
-      itemBuilder: (ctx, index) {
-        widget.nameController.text = customersList[index].name.toString();
-        widget.phoneController.text = customersList[index].phone.toString();
+      itemBuilder: (context, index) {
         CustomerEntity customerList = customersList[index];
         print("object::customerId: ${customerList.customerId.toString()}");
         return GestureDetector(
-          onDoubleTap: () {
-            _deleteCustomer(context, customerList);
-          },
-          onLongPress: () {
-
-            _updateCustomer(context, customerList);
-          },
+          onDoubleTap: () {},
+          onLongPress: () {},
           onTap: () {
-            print("object::customerId: ${customerList.customerId.toString()}");
-            Navigator.push(context, MaterialPageRoute(builder: (_) {
+            Navigator.push(context, MaterialPageRoute(builder: (context) {
               return CustomerDetailPage(
                 customer: customerList,
               );
             }));
           },
           child: ListTile(
+            leading: Icon(Icons.person),
+            trailing: PopupMenuButton<ActionItems>(
+              initialValue: selectedMenu,
+              onSelected: (ActionItems item) {
+                if (item == ActionItems.update) {
+                  _updateCustomer(context, customerList);
+                } else if (item == ActionItems.delete) {
+                  _deleteCustomer(context, customerList);
+                }
+              },
+              itemBuilder: (BuildContext context) =>
+                  <PopupMenuEntry<ActionItems>>[
+                PopupMenuItem<ActionItems>(
+                  value: ActionItems.update,
+                  child: Row(
+                    children: [Icon(Icons.update), Text(" yangilash")],
+                  ),
+                ),
+                PopupMenuItem<ActionItems>(
+                  value: ActionItems.delete,
+                  child: Row(
+                    children: [Icon(Icons.delete), Text(" o\'chirish")],
+                  ),
+                ),
+              ],
+            ),
             title: Text(customerList.name),
             subtitle: Text(customerList.phone),
           ),
         );
       },
-      separatorBuilder: (ctx, index) {
+      separatorBuilder: (context, index) {
         return Divider(
           color: Colors.grey[400],
         );
@@ -79,28 +101,40 @@ class _CustomerListState extends State<CustomerList> {
   void _deleteCustomer(BuildContext context, CustomerEntity customerList) {
     showDialog(
         context: context,
-        builder: (_) {
+        builder: (context) {
           return Center(
               child: AlertDialog(
-            title: Text('Customer delete !'),
-            content: Text('Are you sure delete this customer?'),
+            icon: Icon(Icons.delete_forever),
+            title: Text('Mijozni o\'chirish'),
+            content: Row(
+              children: [
+                Text(
+                  '${customerList.name}',
+                  style:
+                      TextStyle(color: Colors.red, fontWeight: FontWeight.bold),
+                ),
+                Text(' o\'chirilsinmi?'),
+              ],
+            ),
             actions: [
               ElevatedButton(
                   onPressed: () {
-                    setState(() {
-                      context
-                          .read<CustomerCubit>()
-                          .deleteCustomer(customerList.customerId!);
-                    });
-                    Navigator.pop(context);
+                    BlocProvider.of<CustomerCubit>(context, listen: false)
+                        .deleteCustomer(customerList.customerId!);
+
+                    Navigator.pushAndRemoveUntil(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) => CustomersPage()),
+                        (route) => false);
                     print('deleted');
                   },
-                  child: Text('Yes')),
+                  child: Text('Ha')),
               ElevatedButton(
                   onPressed: () {
                     Navigator.pop(context);
                   },
-                  child: Text('No')),
+                  child: Text('Yo\'q')),
             ],
           ));
         });
@@ -108,62 +142,12 @@ class _CustomerListState extends State<CustomerList> {
 
 //
   void _updateCustomer(BuildContext context, CustomerEntity customerList) {
-    showDialog<void>(
-        context: context,
-        builder: (ctx) {
-          return Center(
-              child: AlertDialog(
-            title: Text('Update customer'),
-            content: Column(
-              children: [
-                TextField(
-                  keyboardType: TextInputType.name,
-                  decoration: InputDecoration(
-                    prefixIcon: const Icon(Icons.person),
-                    labelText: 'Name',
-                  ),
-                  controller: widget.nameController,
-                ),
-                TextField(
-                  keyboardType: TextInputType.phone,
-                  decoration: InputDecoration(
-                    prefixIcon: const Icon(Icons.phone),
-                    labelText: 'Phone',
-                  ),
-                  controller: widget.phoneController,
-                ),
-              ],
-            ),
-            actions: [
-              TextButton(
-                style: TextButton.styleFrom(
-                  textStyle: Theme.of(context).textTheme.labelLarge,
-                ),
-                child: const Text('Update'),
-                onPressed: () {
-                  final customerId = customerList.customerId;
-                  CustomerEntity entity = CustomerEntity(
-
-                      name: widget.nameController.text,
-                      phone: widget.phoneController.text,
-                      dateOfSignUp: DateTime.now().millisecondsSinceEpoch,
-                      customerId: customerList.customerId, managerAdded: true);
-
-
-                  setState(() {
-                    context
-                        .read<CustomerCubit>()
-                        .updateCustomer(entity, customerId!);
-                    widget.nameController.clear();
-                    widget.phoneController.clear();
-                  });
-                  Navigator.of(context).pop();
-                },
-              ),
-            ],
-          ));
-        });
+    Navigator.push(
+        context,
+        MaterialPageRoute(
+            builder: (context) => CustomerUpdatePage(
+                  customerList: customerList,
+                )));
   }
 //
-
 }
